+++ src/tree_store/mod.rs
@@ -22,7 +22,7 @@ pub(crate) use page_store::ReadOnlyBackend;
 pub(crate) use page_store::{
     AllocationPolicy, FILE_FORMAT_VERSION3, MAX_PAIR_LENGTH, MAX_VALUE_LENGTH, PAGE_SIZE, Page,
     PageAllocator, PageHint, PageNumber, PageNumberHashSet, PageResolver, PageTrackerPolicy,
-    SerializedSavepoint, ShrinkPolicy, TransactionalMemory,
+    ReadTransactionPageCache, SerializedSavepoint, ShrinkPolicy, TransactionalMemory,
 };
 pub use page_store::{InMemoryBackend, Savepoint, file_backend};
 pub(crate) use table_tree::{PageListMut, TableTree, TableTreeMut};