+++ src/tree_store/page_store/mod.rs
@@ -11,6 +11,7 @@ mod lru_cache;
 mod page_manager;
 mod region;
 mod savepoint;
+mod transaction_page_cache;
 #[allow(clippy::pedantic, dead_code)]
 mod xxh3;
 
@@ -27,6 +28,7 @@ pub(crate) use page_manager::{
 };
 pub use savepoint::Savepoint;
 pub(crate) use savepoint::SerializedSavepoint;
+pub(crate) use transaction_page_cache::ReadTransactionPageCache;
 
 pub(super) use base::{PageImpl, PageMut};
 pub(super) use xxh3::hash128_with_seed;